feature(lowpower): add ulp sleep function
1. add entering deepsleep in ULP code.
2. update chip current when ULP is awake.

+++ examples/lowpower_evb/components/lowpower_evb_ulp/include/lowpower_evb_ulp_opt.h
@@ -40,9 +40,9 @@ void set_ulp_read_value_number(uint16_t value_number);
 /**
  * @brief set interval of ulp read sensor.
  *
- * @value_number value number.
+ * @time_ms time interval(ms).
  */
-void set_ulp_read_interval(uint16_t time_ms);
+void set_ulp_read_interval(uint32_t time_ms);
 
 /**
  * @brief get humidity value from RTC memory.
